Split version parsing from ServiceVersion and LibararyVersion into parent
